refactor: FeedDetail 에서 Post로 변경
