+++ breaking-front/src/pages/Post/units/Comment.js
@@ -4,8 +4,8 @@ import { UserInformationContext } from 'providers/UserInformationProvider';
 import { PAGE_PATH } from 'constants/path';
 import Toggle from 'components/Toggle/Toggle';
 import ProfileImage from 'components/ProfileImage/ProfileImage';
-import CommentForm from 'pages/FeedDetail/units/CommentForm';
-import * as Style from 'pages/FeedDetail/units/Comment.styles';
+import CommentForm from 'pages/Post/units/CommentForm';
+import * as Style from 'pages/Post/units/Comment.styles';
 import { ReactComponent as LikeIcon } from 'assets/svg/like.svg';
 import { ReactComponent as LikedIcon } from 'assets/svg/liked.svg';
 import { ReactComponent as ETCIcon } from 'assets/svg/etc.svg';
